refactor: Update import statements and package structure in ExampleService files
