+++ JAVA/src/main/java/com/mock/bodyguards/service/impl/ExampleServiceImpl.java
@@ -1,7 +1,7 @@
 package com.mock.bodyguards.service.impl;
 
-import com.mock.bodyguards.dto.request.example.ExampleRequest;
-import com.mock.bodyguards.dto.response.example.ExampleResponse;
+import com.mock.bodyguards.dto.example.ExampleRequest;
+import com.mock.bodyguards.dto.example.ExampleResponse;
 import com.mock.bodyguards.entity.Example;
 import com.mock.bodyguards.exception.AlreadyExistsException;
 import com.mock.bodyguards.mapper.ExampleMapper;
